+++ src/atoms/Button/button.stories.js
@@ -2,8 +2,10 @@
 // @flow
 import React, { Fragment } from 'react';
 import { storiesOf } from '@storybook/react';
+import { action } from '@storybook/addon-actions';
 import { Button } from './index';
-import { SignOut, Eye, Refresh } from '../Icon';
+import ButtonCard from './ButtonCard';
+import { SignOut, Eye, Refresh, Download } from '../Icon';
 
 storiesOf('atoms/Button', module)
   .add('standard', () => (
@@ -15,8 +17,29 @@ storiesOf('atoms/Button', module)
   ))
   .add('with icon', () => (
     <Fragment>
-      <Button><Refresh /> Refresh</Button>
-      <Button modifiers={['brand']}><Eye /> Preview</Button>
-      <Button modifiers={['error']}>Sign out <SignOut /></Button>
+      <Button>
+        <Refresh /> Refresh
+      </Button>
+      <Button modifiers={['brand']}>
+        <Eye /> Preview
+      </Button>
+      <Button modifiers={['error']}>
+        Sign out <SignOut />
+      </Button>
+    </Fragment>
+  ))
+  .add('as card', () => (
+    <Fragment>
+      <ButtonCard
+        title="A special file"
+        label="Download"
+        icon={Download}
+        onClick={action('click')}
+      />
+      <ButtonCard
+        title="A very much special file"
+        label="Get over it!"
+        onClick={action('click')}
+      />
     </Fragment>
   ));
